pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "financial-rag"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111",
    "uvicorn[standard]>=0.29",
    "pydantic-settings>=2.0",
    "openai>=1.30",
    "langchain>=0.2",
    "langchain-openai>=0.1",
    "langchain-postgres>=0.0.6",
    "psycopg[binary,pool]>=3.1",
    "pgvector>=0.2",
    "beautifulsoup4>=4.12",
    "pymupdf>=1.24",
    "httpx>=0.27",
    "python-multipart>=0.0.9",
]

[tool.setuptools]
py-modules = ["config", "models", "main"]
